refactor(story-brief): extract schema validation into validation module
